// apps/odonto-front/src/generated/zod/createProcedureDtoSchema.ts
/**
 * Generated by Kubb (https://kubb.dev/).
 * Do not edit manually.
 */

import { z } from "zod/v4";

export const createProcedureDtoSchema = z.object({
  description: z.optional(z.string()),
  type: z.optional(z.string()),
  date: z.string(),
  cost: z.optional(z.number()),
  patientId: z.number(),
  toothNumber: z.optional(z.string()),
  toothFaces: z.optional(z.string()),
});